requirements.txt
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1s
whitenoise==5.1.0
django-heroku==0.3.1
django-rest-framework-social-oauth2==1.1.0
twilio==6.44.0